Finish moving code around
