+++ src/flask_openapi/__init__.py
@@ -6,7 +6,7 @@ from flask_openapi.core import OpenAPI  # noqa
 from flask_openapi.core import Swagger  # noqa
 from flask_openapi.core import Flasgger  # noqa
 from flask_openapi.utils import apispec_to_template  # noqa
-from flask_openapi.utils.decorator import swag_from  # noqa
+from flask_openapi.utils.decorators import swag_from  # noqa
 from flask_openapi.utils.encoder import LazyJSONEncoder  # noqa
 from flask_openapi.utils.sanitizers import BR_SANITIZER  # noqa
 from flask_openapi.utils.sanitizers import MK_SANITIZER  # noqa
